auth-service/requirements.txt
alembic==1.16.4
asyncpg==0.30.0
click==8.2.1
ecdsa==0.19.1
fastapi==0.104.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.6
pyasn1==0.6.1
pydantic==2.5.0
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.6
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.14.1
uvicorn==0.22.0
pydantic-settings==2.1.0
httpx==0.25.2
